fix(contacts): fold the import into one file and one mount point
Review feedback on #342.

The dialog told the user both "please do not close this tab" and "you
can close this dialog", and its running state duplicated the progress
snackbar. Starting an import now closes the dialog, so the snackbar is
the single place progress is reported. The dialog keeps only its idle
state (sample, file input, submit), and the running, complete and error
blocks are gone along with the import.running and
import.background_hint messages. The remaining time moves to the
snackbar, where the progress it belongs to now lives. The import button
is disabled while an import runs, since starting a second one aborted
the first through importIdRef.

withContactImport, ContactImportContext, ContactImportDialog and
ContactImportProgressToast are gone. The context, the provider, the
dialog and the snackbar live in ContactImportProvider.tsx, which is the
only file the registry gains, and Layout and MobileLayout mount the
provider directly. An app passing its own layout to CRM mounts the
provider itself, as the error thrown by useContactImportContext asks;
StoryWrapper does exactly that. CRM.tsx is back to its state on main,
without the wrapped layout and the two useMemo it needed.

Tests start the import through the file input instead of a
startImport() call exposed on the context for their benefit, which
covers the dialog closing itself, and cover the disabled button while
an import runs.
