+++ src/components/atomic-crm/contacts/ContactImportButton.tsx
@@ -2,16 +2,22 @@ import { Upload } from "lucide-react";
 import { useTranslate } from "ra-core";
 import { Button } from "@/components/ui/button";
 
-import { useContactImportContext } from "./ContactImportContext";
+import { useContactImportContext } from "./ContactImportProvider";
 
 export const ContactImportButton = () => {
   const translate = useTranslate();
-  const { openDialog } = useContactImportContext();
+  const { isImporting, openDialog } = useContactImportContext();
 
   return (
     <Button
       variant="outline"
       onClick={openDialog}
+      disabled={isImporting}
+      title={
+        isImporting
+          ? translate("resources.contacts.import.in_progress")
+          : undefined
+      }
       className="flex items-center gap-2 cursor-pointer"
     >
       <Upload /> {translate("resources.contacts.import.button")}
