--- src/components/atomic-crm/contacts/withContactImport.tsx
@@ -1,11 +1,0 @@
-import type { LayoutComponent } from "ra-core";
-
-import { ContactImportProvider } from "./ContactImportProvider";
-
-export const withContactImport =
-  (AppLayout: LayoutComponent): LayoutComponent =>
-  ({ children }) => (
-    <AppLayout>
-      <ContactImportProvider>{children}</ContactImportProvider>
-    </AppLayout>
-  );
