+++ src/components/atomic-crm/layout/Layout.tsx
@@ -4,13 +4,14 @@ import { Notification } from "@/components/admin/notification";
 import { Error } from "@/components/admin/error";
 import { Skeleton } from "@/components/ui/skeleton";
 
+import { ContactImportProvider } from "../contacts/ContactImportProvider";
 import { useConfigurationLoader } from "../root/useConfigurationLoader";
 import Header from "./Header";
 
 export const Layout = ({ children }: { children: ReactNode }) => {
   useConfigurationLoader();
   return (
-    <>
+    <ContactImportProvider>
       <Header />
       <main className="max-w-screen-xl mx-auto pt-4 px-4" id="main-content">
         <ErrorBoundary FallbackComponent={Error}>
@@ -20,6 +21,6 @@ export const Layout = ({ children }: { children: ReactNode }) => {
         </ErrorBoundary>
       </main>
       <Notification />
-    </>
+    </ContactImportProvider>
   );
 };
