+++ src/components/atomic-crm/providers/commons/frenchCrmMessages.ts
@@ -122,10 +122,7 @@ export const frenchCrmMessages = {
           "Échec de l'importation de ce fichier. Veuillez vous assurer que vous avez fourni un fichier CSV valide.",
         imported: "Importé",
         in_progress: "Import des contacts en cours…",
-        background_hint:
-          "Vous pouvez fermer cette fenêtre : l'import continue et vous serez notifié à la fin.",
         remaining_time: "Temps restant estimé :",
-        running: "L'import est en cours, merci de ne pas fermer cet onglet.",
         sample_download: "Télécharger un exemple CSV",
         sample_hint:
           "Voici un exemple de fichier CSV que vous pouvez utiliser comme modèle",
