+++ src/test/StoryWrapper.tsx
@@ -8,6 +8,7 @@ import { createDataProvider } from "@/components/atomic-crm/providers/fakerest";
 import { DEFAULT_USER } from "@/components/atomic-crm/providers/fakerest/authProvider";
 import type { Db } from "@/components/atomic-crm/providers/fakerest/dataGenerator/types";
 import type { Contact, Sale } from "@/components/atomic-crm/types";
+import { ContactImportProvider } from "@/components/atomic-crm/contacts/ContactImportProvider";
 import { CRM } from "@/components/atomic-crm/root/CRM";
 import { testI18nProvider } from "@/components/atomic-crm/providers/commons/i18nProvider";
 
@@ -114,10 +115,10 @@ export const StoryWrapper = ({
         store={store}
         disableTelemetry
         layout={({ children }) => (
-          <>
+          <ContactImportProvider>
             {children}
             <Notification />
-          </>
+          </ContactImportProvider>
         )}
       />
     </MemoryRouter>
